Added DB functionality.
Added code to connect to a database and collect watthr data. Each
collection is associated with a device name (stereo, microwave, etc)
which must already exist in the db. Currently no functionality in
wattcher.py to add new devices to the db. Might add that as a
separate script later. power.sql is all the table creation SQL which
can be read into a MySQL db to create the tables.